backend/pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "circuit_seq_server"
description = "Add short description here"
readme = "README.md"
maintainers = [{ name = "Liam Keegan", email = "ssc@iwr.uni-heidelberg.de" }]
dynamic = ["version"]
requires-python = ">=3.8"
license = { text = "MIT" }
classifiers = [
  "Programming Language :: Python :: 3",
  "Operating System :: OS Independent",
  "License :: OSI Approved :: MIT License",
]
dependencies = [
  "flask",
  "flask-cors",
  "flask-jwt-extended",
  "flask-sqlalchemy",
  "argon2-cffi",
  "click",
  "biopython",
  "gunicorn",
  "snapgene_reader",
  "itsdangerous",
]

[project.scripts]
circuit_seq_server = "circuit_seq_server.main:main"

[project.optional-dependencies]
tests = ["pytest", "pytest-cov", "freezegun"]
docs = ["m2r2", "sphinx", "sphinx_rtd_theme"]

[tool.setuptools.dynamic]
version = { attr = "circuit_seq_server.__version__" }

[tool.pytest.ini_options]
testpaths = ["tests"]
